# Conflicts:
#	services/headless-lms/models/.sqlx/query-4ad4fa5a7945903b7dd147844b61e816b671c117197da3f60bb4560fb09cee62.json

+++ services/headless-lms/utils/src/services/sisu.rs
@@ -143,8 +143,8 @@ impl Additional {
 
 #[derive(Serialize, Deserialize, ToSchema, Debug)]
 pub struct Credits {
-    pub min: Option<i64>,
-    pub max: Option<i64>,
+    pub min: Option<f32>,
+    pub max: Option<f32>,
 }
 
 #[derive(Serialize, Deserialize, ToSchema, Debug)]
@@ -160,7 +160,7 @@ pub struct Organisation {
     pub organisation_id: Option<String>,
     pub educational_institution_urn: Option<serde_json::Value>,
     pub role_urn: String,
-    pub share: i64,
+    pub share: f64,
     pub validity_period: Option<OrganisationValidityPeriod>,
 }
